recognition/requirements.txt
tensorflow
opencv-python
tqdm
requests
pandas
numpy<2.0
pillow